parkHereWebApp/requirements.txt
anyio==3.7.1
APScheduler==3.11.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
dotenv==0.9.9
exceptiongroup==1.2.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
pyproj==3.7.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
